Add inital version of a image processing based solution
This is a basic image processing based solution, detecting green
pixels. Based on the detection of the color the pixel is
classifed. The process overview is:

1. Detect pixels with dominant green channel
2. Clip the histogramm to the lower existing values and spread it
   out agian to use all values
3. Make the grayscale image just black/white

For better results a second approch preprocesses the image by
removing shadows, however the current process is artifact heavy
and does not meanifully improve the result.

To better visualize the results edge detection was run on the
masks, withe areas were pink tinted, black areas removed and this
then overlayed onto the original.