Automated and expaned image processing approch
The whole process is now automated, config values however are
still off and need to be ajusted. Further noise reduction is
applied additional. The reduction is based on a gaussian blurr
befor green detection and a morphological close and open on the
created mask to remove small dots in either region.