Change evauation to a python package with proper interface
The evaluation code neede a proper interface in regards to
parameters and return of result values.

The new interface takes either a list of in memeory images
or files as tupels that should be compared. There is also a
single pair evaluation available. The return is handled by a
Data class to give all name a proper autocompleting name and
attached documentation. The old print out still exists as a
function that prints out the evaluation result object as
befor.

To be able to conver to a package the name of the folder had
to be changed from '-' to '_'. This was necessary as Python
does not like '-' in package names.