Parameterize image-processing and add optimiter
With this the image processing approch follows the default
methode call convention for inputs and outputs. Now all methodes
accept a string or in memory image, write to disc if output is
not None and return an in memory image.

All parts of the process can no be configured by function paramters.
The default values of the parmeters are the old values.

Additional a hsv range based detector is now also available.

The parameter are defined with their values and limits in
configurator.py.

To optimize the new parameters the configurator.py provides the
ablility to cross over to parameter sets with an additiona ranom
mutation chance. This is then used by a genetic algorithem in
optimize.py. The algorithem holds n parameter configurations im
memory and evaluates each with the given examples. The better half
will survive to the next round. The other halfe is replaces with
random crossovers of the survivors. This should finde an pretty
good combination of paramters in due time.