Fix exeption in cut_mask_from_image if mask and image mixed grayscale and rgb
During integration in the image-processing approche, the cut_mask_from_image raised exeptions. This was,
because sometimes the image and mask had mixed grayscale and rgb formats.

To fix this, first it is checked if the total numer of elements is larger or smaller in one of them. If the
mask has to few, it is grayscale and shuld be rgb so the one channel is copied 3 times. In the ohter cases
the mask has is rgb and the imrage only grayscale, so any one channel of the mask is good enough.