Improve image processing approche and get ready to run
Multiple improvements are made.

- The genetic algorithem now only evaluates the new configurations, reusing the old ones. This is a ~2x speed up

- The genetic algorithem now supports custom intermediate results files

- The genetic algorithen now supports external seeding of the first popolation

- The genetic algorithem now supports prepreared in memeory images to speed up calculation

- Added saturation and contrast to the parametrs and pipeleine.

- Added house removeal at the beginning and end.

For the first training round the shadow removal steps will not be optimized and precomputed to save time.
HSV detection is currently disable to improve convergence speed.

The results of a first 30 generations evolution are in current_bet_01_07.txt stored and will be used as
additional seed paramers in future runs.